+++ src/test/scala/com/example/presentation/Credentials.scala
@@ -1,0 +1,4 @@
+package com.example
+package presentation
+
+trait Credentials
