+++ src/test/scala/com/example/presentation/package.scala
@@ -1,5 +1,11 @@
 package com.example
 
+/**
+ * Order of examples:
+ * 1. FreeExample
+ * 2. IxStateExample
+ * 3. IxFreeExample
+ */
 package object presentation {
 
   val myCredentials: Credentials = new Credentials {}
